Metodos de precios, paradas y administración

+++ App-Monopatin-Electrico/microViaje/src/main/java/org/example/microviaje/DTO/ViajeRequestDTO.java
@@ -65,7 +65,7 @@ public class ViajeRequestDTO {
 
     @NotNull(message = "El campo precioTotal es obligatorio.")
     @NotEmpty(message = "El campo precioTotal no  puede estar vacío")
-    private double precioTotal;
+    private Float precioTotal;
